bugfix: fixing routes
