+++ .vercel/output/_functions/entry.mjs
@@ -1,6 +1,6 @@
 import { renderers } from './renderers.mjs';
-import { c as createExports, s as serverEntrypointModule } from './chunks/_@astrojs-ssr-adapter_DShy0iUI.mjs';
-import { manifest } from './manifest_BsEgltPs.mjs';
+import { c as createExports, s as serverEntrypointModule } from './chunks/_@astrojs-ssr-adapter_DFZuJGpU.mjs';
+import { manifest } from './manifest_BNYNINxQ.mjs';
 
 const serverIslandMap = new Map();;
 
@@ -43,7 +43,7 @@ const _manifest = Object.assign(manifest, {
     middleware: () => import('./_noop-middleware.mjs')
 });
 const _args = {
-    "middlewareSecret": "05d2b059-6be1-4c9d-ab7a-288cd858c1d1",
+    "middlewareSecret": "1fabeb4d-5738-4406-9698-ab14d537dcce",
     "skewProtection": false
 };
 const _exports = createExports(_manifest, _args);
