+++ .vercel/output/_functions/pages/api/contact.astro.mjs
@@ -30,7 +30,7 @@ async function POST({ request }) {
       "KOMMO_FUENTE_LEAD_ID",
       "KOMMO_FUENTE_LEAD_ENUM_ID"
     ];
-    const missingVars = requiredEnvVars.filter((varName) => !Object.assign(__vite_import_meta_env__, { KOMMO_SUBDOMAIN: "colinasdelapaz", KOMMO_ACCESS_TOKEN: "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6Ijc0YzBjOWQ0YTA2MGY2NzUyYWE0MzVlMGQ0NTk5NjczMDM3MTliYzgyMjJhYjAxOTc0ZmNlNjAyYTZiODhmZWE3OWUwNTEzYzE2ODFhMGEzIn0.eyJhdWQiOiI4MjM0Yjc2Yy1lYTMyLTRmNmYtYjU4Zi01MGUwYjAzZDI5MDQiLCJqdGkiOiI3NGMwYzlkNGEwNjBmNjc1MmFhNDM1ZTBkNDU5OTY3MzAzNzE5YmM4MjIyYWIwMTk3NGZjZTYwMmE2Yjg4ZmVhNzllMDUxM2MxNjgxYTBhMyIsImlhdCI6MTc1NjkxNjMwNSwibmJmIjoxNzU2OTE2MzA1LCJleHAiOjE4OTMzNjk2MDAsInN1YiI6IjEyNTc1MjgzIiwiZ3JhbnRfdHlwZSI6IiIsImFjY291bnRfaWQiOjM0MDQ5NzQzLCJiYXNlX2RvbWFpbiI6ImtvbW1vLmNvbSIsInZlcnNpb24iOjIsInNjb3BlcyI6WyJjcm0iLCJmaWxlcyIsImZpbGVzX2RlbGV0ZSIsIm5vdGlmaWNhdGlvbnMiLCJwdXNoX25vdGlmaWNhdGlvbnMiXSwidXNlcl9mbGFncyI6MCwiaGFzaF91dWlkIjoiYWU3M2MyZGUtZGRhNS00MWZhLWExNzAtOGVkYjQ1ZDlmZWQxIiwiYXBpX2RvbWFpbiI6ImFwaS1jLmtvbW1vLmNvbSJ9.KIhzWazN8QOJgE53KCX_zJTHDhLmEE_Znc92XuoPZHXfSnul2tsNbqAfF8YiuxWS9RUj0sBQKIScQ0Wv2jeQs7rbEjO3VeoTt2QQAcvSG2ZUkA3JS456uuBMpSuMNs1VkFf8rTqP2PbcKquGdKwhlbSAYCXPyw1G5WSeT5YF5LXfSj3ZCPp1bU37tiw3actyLVOfDOPYVxSC3zm5e973CScSl2a-zp-emllag5TPa43f5vdR0wYwmBxDcePyGK6OtQVk400Br41c__0oknbc42DsPJENQ9_G3jmCuSVtRcpbJEJLnh715-RIhsLeWpbRxcGbmS6gOfNlSVxbdkpQNg", KOMMO_PIPELINE_ID: "10348799", KOMMO_USER_ID: "13774007", KOMMO_EMAIL_FIELD_ID: "119884", KOMMO_PHONE_FIELD_ID: "119882", KOMMO_SERVICE_FIELD_ID: "1021440", KOMMO_FUENTE_LEAD_ID: "1021438", KOMMO_FUENTE_LEAD_ENUM_ID: "974566", OS: process.env.OS })[varName]);
+    const missingVars = requiredEnvVars.filter((varName) => !Object.assign(__vite_import_meta_env__, { KOMMO_SUBDOMAIN: "colinasdelapaz", KOMMO_ACCESS_TOKEN: "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6Ijc0YzBjOWQ0YTA2MGY2NzUyYWE0MzVlMGQ0NTk5NjczMDM3MTliYzgyMjJhYjAxOTc0ZmNlNjAyYTZiODhmZWE3OWUwNTEzYzE2ODFhMGEzIn0.eyJhdWQiOiI4MjM0Yjc2Yy1lYTMyLTRmNmYtYjU4Zi01MGUwYjAzZDI5MDQiLCJqdGkiOiI3NGMwYzlkNGEwNjBmNjc1MmFhNDM1ZTBkNDU5OTY3MzAzNzE5YmM4MjIyYWIwMTk3NGZjZTYwMmE2Yjg4ZmVhNzllMDUxM2MxNjgxYTBhMyIsImlhdCI6MTc1NjkxNjMwNSwibmJmIjoxNzU2OTE2MzA1LCJleHAiOjE4OTMzNjk2MDAsInN1YiI6IjEyNTc1MjgzIiwiZ3JhbnRfdHlwZSI6IiIsImFjY291bnRfaWQiOjM0MDQ5NzQzLCJiYXNlX2RvbWFpbiI6ImtvbW1vLmNvbSIsInZlcnNpb24iOjIsInNjb3BlcyI6WyJjcm0iLCJmaWxlcyIsImZpbGVzX2RlbGV0ZSIsIm5vdGlmaWNhdGlvbnMiLCJwdXNoX25vdGlmaWNhdGlvbnMiXSwidXNlcl9mbGFncyI6MCwiaGFzaF91dWlkIjoiYWU3M2MyZGUtZGRhNS00MWZhLWExNzAtOGVkYjQ1ZDlmZWQxIiwiYXBpX2RvbWFpbiI6ImFwaS1jLmtvbW1vLmNvbSJ9.KIhzWazN8QOJgE53KCX_zJTHDhLmEE_Znc92XuoPZHXfSnul2tsNbqAfF8YiuxWS9RUj0sBQKIScQ0Wv2jeQs7rbEjO3VeoTt2QQAcvSG2ZUkA3JS456uuBMpSuMNs1VkFf8rTqP2PbcKquGdKwhlbSAYCXPyw1G5WSeT5YF5LXfSj3ZCPp1bU37tiw3actyLVOfDOPYVxSC3zm5e973CScSl2a-zp-emllag5TPa43f5vdR0wYwmBxDcePyGK6OtQVk400Br41c__0oknbc42DsPJENQ9_G3jmCuSVtRcpbJEJLnh715-RIhsLeWpbRxcGbmS6gOfNlSVxbdkpQNg", KOMMO_PIPELINE_ID: "10348799", KOMMO_USER_ID: "13774007", KOMMO_EMAIL_FIELD_ID: "119884", KOMMO_PHONE_FIELD_ID: "119882", KOMMO_SERVICE_FIELD_ID: "1021440", KOMMO_FUENTE_LEAD_ID: "1021438", KOMMO_FUENTE_LEAD_ENUM_ID: "974566", OS: process.env.OS, _: process.env._ })[varName]);
     if (missingVars.length > 0) {
       console.error("Variables de entorno faltantes:", missingVars);
       return new Response(JSON.stringify({
