+++ .vercel/output/_functions/pages/faq.astro.mjs
@@ -1,48 +1,1 @@
-import { f as createComponent, r as renderTemplate, k as renderComponent, m as maybeRenderHead } from '../chunks/astro/server_fVTS04Ou.mjs';
-import 'kleur/colors';
-import { $ as $$Layout } from '../chunks/Layout_W9iYiw_y.mjs';
-export { renderers } from '../renderers.mjs';
-
-var __freeze = Object.freeze;
-var __defProp = Object.defineProperty;
-var __template = (cooked, raw) => __freeze(__defProp(cooked, "raw", { value: __freeze(raw || cooked.slice()) }));
-var _a;
-const $$Faq = createComponent(($$result, $$props, $$slots) => {
-  return renderTemplate(_a || (_a = __template(["", " <script>\n  function toggleFaq(id) {\n    const content = document.getElementById(`content-${id}`);\n    const icon = document.getElementById(`icon-${id}`);\n    \n    if (content && icon) {\n      if (content.classList.contains('hidden')) {\n        content.classList.remove('hidden');\n        icon.style.transform = 'rotate(180deg)';\n      } else {\n        content.classList.add('hidden');\n        icon.style.transform = 'rotate(0deg)';\n      }\n    }\n  }\n<\/script>"], ["", " <script>\n  function toggleFaq(id) {\n    const content = document.getElementById(\\`content-\\${id}\\`);\n    const icon = document.getElementById(\\`icon-\\${id}\\`);\n    \n    if (content && icon) {\n      if (content.classList.contains('hidden')) {\n        content.classList.remove('hidden');\n        icon.style.transform = 'rotate(180deg)';\n      } else {\n        content.classList.add('hidden');\n        icon.style.transform = 'rotate(0deg)';\n      }\n    }\n  }\n<\/script>"])), renderComponent($$result, "Layout", $$Layout, { "title": "Preguntas Frecuentes - Colinas de la Paz" }, { "default": ($$result2) => renderTemplate` ${maybeRenderHead()}<section class="py-16 bg-white"> <div class="container mx-auto px-6"> <!-- Title --> <div class="text-center mb-12"> <h1 class="text-4xl lg:text-5xl font-bold text-dark-gray mb-4">
-Preguntas Frecuentes
-</h1> </div> <!-- FAQ Accordion --> <div class="max-w-4xl mx-auto space-y-4"> <!-- FAQ 1 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(1)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-1. ¿Qué debo hacer al momento de la pérdida de un ser querido?
-</h3> <svg id="icon-1" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-1" class="hidden px-6 pb-4"> <div class="text-dark-gray/80 space-y-3"> <p><strong>Si ocurre una muerte natural (edad 60 + o enfermedad crónica):</strong> Contacta de inmediato a un médico o institución autorizada, quienes harán el dictamen correspondiente. Una vez confirmado oficialmente, puedes notificarnos para gestionar el procedimiento.</p> <p><strong>Si fallece por accidente o causa externa:</strong> Reporta el suceso a la Fiscalía o entidad judicial competente para su investigación.</p> <p><strong>Si hubo emergencia médica al instante:</strong> Llama al 911 o número local de emergencias para asistencia y certificación médica.</p> </div> </div> </div> <!-- FAQ 2 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(2)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-2. ¿Dónde está ubicado el Campo Santo Colinas de la Paz?
-</h3> <svg id="icon-2" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-2" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>El Camposanto está ubicado en la República de Panamá, provincia de Panamá Oeste, corregimiento Juan Demóstenes Arosemena, Hato Montaña Arraiján.</p> </div> </div> </div> <!-- FAQ 3 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(3)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-3. ¿Además del Campo Santo, donde puedo preguntar por planes y servicios?
-</h3> <svg id="icon-3" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-3" class="hidden px-6 pb-4"> <div class="text-dark-gray/80 space-y-3"> <p>Colinas de la Paz tiene su oficina dentro del Campo Santo. El horario de oficina es de:</p> <div class="ml-4"> <p>Lunes a viernes: 8:00 am a 5:00 pm</p> <p>Sábado: 8:00 am a 3:00 pm</p> <p>Domingo: Cerrado</p> </div> <p>También puedes escribirnos a través de nuestras redes sociales, correo electrónico a ventas@colinasdelapaz.com o llenar el formulario de contacto en la sección Agenda tu visita.</p> </div> </div> </div> <!-- FAQ 4 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(4)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-4. ¿Cuál es la diferencia entre Pre-Necesidad y Necesidad Inmediata?
-</h3> <svg id="icon-4" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-4" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>Pre-Necesidad o planificación anticipada es la mejor alternativa vitalicia para adquirir una protección exequial de forma anticipada en cómodas cuotas que permiten una flexibilidad económica para quien la contrata; mientras que en la necesidad inmediata usted compra el bien o servicio en el momento de la pérdida de un ser querido.</p> </div> </div> </div> <!-- FAQ 6 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(6)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-6. ¿Cuál es el precio de un Servicio Funerario?
-</h3> <svg id="icon-6" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-6" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>Los precios varían. Para mayor información, puede escribir al whatsapp +507 6675-3132.</p> </div> </div> </div> <!-- FAQ 7 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(7)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-7. ¿Se puede utilizar el lote antes de cancelar la totalidad del pago?
-</h3> <svg id="icon-7" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-7" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>Si, puede utilizar el lote siempre y cuando se cancele un porcentaje igual o mayor al 60% del precio total estipulado en el contrato.</p> </div> </div> </div> <!-- FAQ 8 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(8)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-8. ¿Cómo puedo realizar el pago de mi plan mensual?
-</h3> <svg id="icon-8" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-8" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>Puede realizar sus pagos mediante banca en línea en pagos a terceros, Yappy, descuento directo con su tarjeta de crédito, en la oficina del Camposanto, Punto Pago y en Multipagos.</p> </div> </div> </div> <!-- FAQ 9 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(9)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-9. ¿Cuáles son los horarios del Campo Santo?
-</h3> <svg id="icon-9" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-9" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>Puede visitar el Camposanto de 6:00am a 6:00pm.</p> </div> </div> </div> <!-- FAQ 10 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(10)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-10. ¿Puede variar el precio después de haber adquirido un servicio?
-</h3> <svg id="icon-10" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-10" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>No, al haber adquirido un servicio por un precio estipulado, este se mantiene hasta finalizar el contrato en que se adquirió.</p> </div> </div> </div> <!-- FAQ 11 --> <div class="bg-light-gray/20 rounded-lg shadow-sm border border-light-gray/30"> <button class="w-full px-6 py-4 text-left flex justify-between items-center hover:bg-gray-50 transition-colors" onclick="toggleFaq(11)"> <h3 class="text-lg font-semibold text-dark-gray pr-4">
-11. ¿Qué sucede en el caso de que el dueño del contrato fallezca?
-</h3> <svg id="icon-11" class="w-5 h-5 text-dark-gray transform transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div id="content-11" class="hidden px-6 pb-4"> <div class="text-dark-gray/80"> <p>En caso de que el dueño fallezca, los derechos del contrato pasan directamente al primer encargado.</p> </div> </div> </div> </div> </div> </section> ` }));
-}, "C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/faq.astro", void 0);
-
-const $$file = "C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/faq.astro";
-const $$url = "/faq";
-
-const _page = /*#__PURE__*/Object.freeze(/*#__PURE__*/Object.defineProperty({
-  __proto__: null,
-  default: $$Faq,
-  file: $$file,
-  url: $$url
-}, Symbol.toStringTag, { value: 'Module' }));
-
-const page = () => _page;
-
-export { page };
+// Contents removed by Astro as it's used for prerendering only
