--- .vercel/output/functions/_render.func/src/pages/aboutus.astro
@@ -1,15 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/aboutus/Hero.astro';
-import Description from '../components/aboutus/Description.astro';
-import CTA from '../components/aboutus/CTA.astro';
----
-
-<Layout 
-  title="Sobre Nosotros - Colinas de la Paz | Historia y Experiencia"
-  description="Conoce la historia de Colinas de la Paz. Más de 27 años de experiencia en servicios funerarios en Panamá Oeste. Misión, visión y valores que nos guían."
->
-  <Hero />
-  <Description />
-  <CTA />
-</Layout>
