--- .vercel/output/functions/_render.func/src/pages/burning.astro
@@ -1,19 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/services/burning/Hero.astro';
-import Alternative from '../components/services/burning/Alternative.astro';
-import UrnsVariety from '../components/services/burning/UrnsVariety.astro';
-import Osarios from '../components/services/burning/Osarios.astro';
-import CTA from '../components/services/burning/CTA.astro';
----
-
-<Layout 
-  title="Servicio de Cremación - Colinas de la Paz | Nichos, Cremaciones y Osarios"
-  description="Servicio de cremación en Colinas de la Paz. Nichos, cremaciones y osarios, variedad de urnas y ceremonias religiosas. Servicio completo con traslado y trámites legales incluidos."
->
-  <Hero />
-  <Alternative />
-  <UrnsVariety />
-  <Osarios />
-  <CTA />
-</Layout>
