--- .vercel/output/functions/_render.func/src/pages/funservices.astro
@@ -1,17 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/services/funservices/Hero.astro';
-import Opportunity from '../components/services/funservices/Opportunity.astro';
-import ServicesOffered from '../components/services/funservices/ServicesOffered.astro';
-import CTA from '../components/services/funservices/CTA.astro';
----
-
-<Layout 
-  title="Servicios Funerarios - Colinas de la Paz | Servicios Completos"
-  description="Servicios funerarios completos en Colinas de la Paz. Ceremonias, velatorios, traslados y trámites. Acompañamiento profesional en momentos difíciles."
->
-  <Hero />
-  <Opportunity />
-  <ServicesOffered />
-  <CTA />
-</Layout>
