--- .vercel/output/functions/_render.func/src/pages/optional.astro
@@ -1,17 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/services/optional/Hero.astro';
-import Description from '../components/services/optional/Description.astro';
-import ServiceDetails from '../components/services/optional/ServiceDetails.astro';
-import CTA from '../components/services/optional/CTA.astro';
----
-
-<Layout 
-  title="Servicios Opcionales - Colinas de la Paz | Salas de Velación"
-  description="Servicios opcionales en Colinas de la Paz. Salas de velación con aire acondicionado, equipo de sonido, cafetera y responsable 24 horas. Comodidad para las familias."
->
-  <Hero />
-  <Description />
-  <ServiceDetails />
-  <CTA />
-</Layout>
