--- .vercel/output/functions/_render.func/src/pages/ponline.astro
@@ -1,10 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/ponline/Hero.astro';
-import PaymentMethods from '../components/ponline/PaymentMethods.astro';
----
-
-<Layout title="Pagos en Línea - Colinas de la Paz">
-  <Hero />
-  <PaymentMethods />
-</Layout>
