--- .vercel/output/functions/_render.func/src/pages/preneed.astro
@@ -1,14 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/Preneed/Hero.astro';
-import Description from '../components/Preneed/Description.astro';
-import Planning from '../components/Preneed/Planning.astro';
-import CTA from '../components/Preneed/CTA.astro';
----
-
-<Layout>
-  <Hero />
-  <Description />
-  <Planning />
-  <CTA />
-</Layout>
