--- .vercel/output/functions/_render.func/src/pages/slots.astro
@@ -1,17 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/services/slots/Hero.astro';
-import Description from '../components/services/slots/Description.astro';
-import Locations from '../components/services/slots/Locations.astro';
-import CTA from '../components/services/slots/CTA.astro';
----
-
-<Layout 
-  title="Lotes - Colinas de la Paz | Espacios Dignos para el Descanso Eterno"
-  description="Lotes individuales y familiares en Colinas de la Paz, Panamá Oeste. Espacios tranquilos y dignos para la sepultura con diferentes sectores disponibles."
->
-  <Hero />
-  <Description />
-  <Locations />
-  <CTA />
-</Layout>
