+++ src/pages/aboutus.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/aboutus/Hero.astro';
 import Description from '../components/aboutus/Description.astro';
