+++ src/pages/burning.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/services/burning/Hero.astro';
 import Alternative from '../components/services/burning/Alternative.astro';
