+++ src/pages/funservices.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/services/funservices/Hero.astro';
 import Opportunity from '../components/services/funservices/Opportunity.astro';
