+++ src/pages/optional.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/services/optional/Hero.astro';
 import Description from '../components/services/optional/Description.astro';
