+++ src/pages/preneed.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/Preneed/Hero.astro';
 import Description from '../components/Preneed/Description.astro';
@@ -6,7 +8,10 @@ import Planning from '../components/Preneed/Planning.astro';
 import CTA from '../components/Preneed/CTA.astro';
 ---
 
-<Layout>
+<Layout 
+  title="Prenecesidad - Colinas de la Paz | Planificación Funeraria"
+  description="Servicios de prenecesidad en Colinas de la Paz. Planifica con anticipación y asegura tranquilidad para tu familia. Opciones de pago flexibles y asesoría especializada."
+>
   <Hero />
   <Description />
   <Planning />
