+++ src/pages/slots.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/services/slots/Hero.astro';
 import Description from '../components/services/slots/Description.astro';
